+++ src/async_client.rs
@@ -1,29 +1,50 @@
-use futures::{future, stream, Future, Stream};
+use futures::prelude::*;
+use futures::{
+    future::{try_join_all, BoxFuture, FutureExt, TryFutureExt},
+    stream::{self, TryStreamExt},
+    try_join,
+};
 use log::trace;
-use reqwest::{header, r#async, StatusCode, Url};
+use reqwest::{header, Client as HttpClient, StatusCode, Url};
 use serde::de::DeserializeOwned;
+
+use std::iter::FromIterator;
 
 use super::Error;
 use crate::types::*;
 
+pub type Result<T> = std::result::Result<T, Error>;
+
 /// Asynchronous client for the crates.io API.
 #[derive(Clone)]
 pub struct Client {
-    client: r#async::Client,
+    client: HttpClient,
     base_url: Url,
 }
 
 impl Client {
     /// Instantiate a new client.
     ///
-    /// This will fail if the underlying http client could not be created.
+    /// This will panic if the underlying http client could not be created.
+    ///
+    /// The user agent will default to `crates_io_api/{crates_io_api version}`.
     pub fn new() -> Self {
+        let mut headers = header::HeaderMap::new();
+        // Crates.io requires a user agent or it will return 403's on all calls
+        headers.insert(
+            header::USER_AGENT,
+            header::HeaderValue::from_static(crate::DEFAULT_USER_AGENT),
+        );
         Self {
-            client: r#async::Client::new(),
+            client: HttpClient::builder()
+                .default_headers(headers)
+                .build()
+                .expect("Could not initialize HTTP client"),
             base_url: Url::parse("https://crates.io/api/v1/").unwrap(),
         }
     }
 
+    /// Create a new client with a specific user agent
     pub fn with_user_agent(user_agent: &str) -> Self {
         let mut headers = header::HeaderMap::new();
         headers.insert(
@@ -31,7 +52,7 @@ impl Client {
             header::HeaderValue::from_str(user_agent).unwrap(),
         );
         Self {
-            client: r#async::Client::builder()
+            client: HttpClient::builder()
                 .default_headers(headers)
                 .build()
                 .unwrap(),
@@ -39,12 +60,13 @@ impl Client {
         }
     }
 
-    fn get<T: DeserializeOwned>(&self, url: &Url) -> impl Future<Item = T, Error = Error> {
+    async fn get<T: DeserializeOwned>(&self, url: &Url) -> Result<T> {
         trace!("GET {}", url);
 
         self.client
             .get(url.clone())
             .send()
+            .await
             .map_err(Error::from)
             .and_then(|res| {
                 if res.status() == StatusCode::NOT_FOUND {
@@ -52,40 +74,42 @@ impl Client {
                 }
                 let res = res.error_for_status()?;
                 Ok(res)
-            })
-            .and_then(|mut res| res.json().map_err(Error::from))
+            })?
+            .json::<T>()
+            .await
+            .map_err(Error::from)
     }
 
     /// Retrieve a summary containing crates.io wide information.
-    pub fn summary(&self) -> impl Future<Item = Summary, Error = Error> {
+    pub async fn summary(&self) -> Result<Summary> {
         let url = self.base_url.join("summary").unwrap();
-        self.get(&url)
+        self.get(&url).await
     }
 
     /// Retrieve information of a crate.
     ///
     /// If you require detailed information, consider using [full_crate]().
-    pub fn get_crate(&self, name: &str) -> impl Future<Item = CrateResponse, Error = Error> {
+    pub async fn get_crate(&self, name: &str) -> Result<CrateResponse> {
         let url = self.base_url.join("crates/").unwrap().join(name).unwrap();
-        self.get(&url)
+        self.get(&url).await
     }
 
     /// Retrieve download stats for a crate.
-    pub fn crate_downloads(&self, name: &str) -> impl Future<Item = Downloads, Error = Error> {
+    pub async fn crate_downloads(&self, name: &str) -> Result<Downloads> {
         let url = self
             .base_url
             .join(&format!("crates/{}/downloads", name))
             .unwrap();
-        self.get(&url)
+        self.get(&url).await
     }
 
     /// Retrieve the owners of a crate.
-    pub fn crate_owners(&self, name: &str) -> impl Future<Item = Vec<User>, Error = Error> {
+    pub async fn crate_owners(&self, name: &str) -> Result<Vec<User>> {
         let url = self
             .base_url
             .join(&format!("crates/{}/owners", name))
             .unwrap();
-        self.get::<Owners>(&url).map(|data| data.users)
+        self.get::<Owners>(&url).await.map(|data| data.users)
     }
 
     /// Load all reverse dependencies of a crate.
@@ -93,68 +117,75 @@ impl Client {
     /// Note: Since the reverse dependency endpoint requires pagination, this
     /// will result in multiple requests if the crate has more than 100 reverse
     /// dependencies.
-    pub fn crate_reverse_dependencies(&self, name: &str)
-        -> impl Future<Item = ReverseDependencies, Error = Error> {
+    pub async fn crate_reverse_dependencies(&self, name: &str) -> Result<ReverseDependencies> {
+        fn fetch_page(
+            c: Client,
+            name: String,
+            mut tidy_rdeps: ReverseDependencies,
+            page: u64,
+        ) -> BoxFuture<'static, Result<ReverseDependencies>> {
+            let url = c
+                .base_url
+                .join(&format!(
+                    "crates/{0}/reverse_dependencies?per_page=100&page={1}",
+                    name, page
+                ))
+                .unwrap();
 
-        fn fetch_page(c: Client, name: String, mut tidy_rdeps: ReverseDependencies, page: u64)
-            -> impl Future<Item = ReverseDependencies, Error = Error> + Send {
-
-            let url = c.base_url.join(&format!(
-                "crates/{0}/reverse_dependencies?per_page=100&page={1}", name, page
-            )).unwrap();
-
-            c.get::<ReverseDependenciesAsReceived>(&url).and_then(move |rdeps|
-                -> Box<dyn Future<Item = ReverseDependencies, Error = Error> + Send> {
-
+            async move {
+                let rdeps = c.get::<ReverseDependenciesAsReceived>(&url).await?;
                 tidy_rdeps.from_received(&rdeps);
 
                 if !rdeps.dependencies.is_empty() {
                     tidy_rdeps.meta = rdeps.meta;
-                    Box::new(fetch_page(c, name, tidy_rdeps, page + 1))
+                    fetch_page(c, name, tidy_rdeps, page + 1).await
                 } else {
-                    Box::new(::futures::future::ok(tidy_rdeps))
+                    Ok(tidy_rdeps)
                 }
-            })
+            }
+            .boxed()
         }
 
-        fetch_page(self.clone(), name.to_string(), ReverseDependencies {
-            dependencies: Vec::new(), meta:Meta{total:0} }, 1)
+        fetch_page(
+            self.clone(),
+            name.to_string(),
+            ReverseDependencies {
+                dependencies: Vec::new(),
+                meta: Meta { total: 0 },
+            },
+            1,
+        )
+        .await
     }
 
     /// Retrieve the authors for a crate version.
-    pub fn crate_authors(
-        &self,
-        name: &str,
-        version: &str,
-    ) -> impl Future<Item = Authors, Error = Error> {
+    pub async fn crate_authors(&self, name: &str, version: &str) -> Result<Authors> {
         let url = self
             .base_url
             .join(&format!("crates/{}/{}/authors", name, version))
             .unwrap();
-        self.get::<AuthorsResponse>(&url).map(|res| Authors {
+        self.get::<AuthorsResponse>(&url).await.map(|res| Authors {
             names: res.meta.names,
             users: res.users,
         })
     }
 
     /// Retrieve the dependencies of a crate version.
-    pub fn crate_dependencies(
-        &self,
-        name: &str,
-        version: &str,
-    ) -> impl Future<Item = Vec<Dependency>, Error = Error> {
+    pub async fn crate_dependencies(&self, name: &str, version: &str) -> Result<Vec<Dependency>> {
         let url = self
             .base_url
             .join(&format!("crates/{}/{}/dependencies", name, version))
             .unwrap();
-        self.get::<Dependencies>(&url).map(|res| res.dependencies)
+        self.get::<Dependencies>(&url)
+            .await
+            .map(|res| res.dependencies)
     }
 
-    fn full_version(&self, version: Version) -> impl Future<Item = FullVersion, Error = Error> {
-        let authors = self.crate_authors(&version.crate_name, &version.num);
-        let deps = self.crate_dependencies(&version.crate_name, &version.num);
+    async fn full_version(&self, version: Version) -> Result<FullVersion> {
+        let authors_fut = self.crate_authors(&version.crate_name, &version.num);
+        let deps_fut = self.crate_dependencies(&version.crate_name, &version.num);
 
-        authors.join(deps).map(|(authors, deps)| FullVersion {
+        try_join!(authors_fut, deps_fut).map(|(authors, deps)| FullVersion {
             created_at: version.created_at,
             updated_at: version.updated_at,
             dl_path: version.dl_path,
@@ -185,70 +216,62 @@ impl Client {
         &self,
         name: &str,
         all_versions: bool,
-    ) -> impl Future<Item = FullCrate, Error = Error> {
+    ) -> impl Future<Output = Result<FullCrate>> {
         let c = self.clone();
-        let crate_and_versions = self.get_crate(name).and_then(
-            move |info| -> Box<
-                Future<Item = (CrateResponse, Vec<FullVersion>), Error = Error> + Send,
-            > {
-                if !all_versions {
-                    Box::new(
-                        c.full_version(info.versions[0].clone())
-                            .map(|v| (info, vec![v])),
-                    )
-                } else {
-                    Box::new(
-                        ::futures::future::join_all(
-                            info.versions
-                                .clone()
-                                .into_iter()
-                                .map(|v| c.full_version(v))
-                                .collect::<Vec<_>>(),
-                        )
-                        .map(|versions| (info, versions)),
-                    )
-                }
-            },
-        );
-
-        let dls = self.crate_downloads(name);
-        let owners = self.crate_owners(name);
-        let reverse_dependencies = self.crate_reverse_dependencies(name);
-
-        crate_and_versions
-            .join4(dls, owners, reverse_dependencies)
-            .map(|((resp, versions), dls, owners, reverse_dependencies)| {
-                let data = resp.crate_data;
-                FullCrate {
-                    id: data.id,
-                    name: data.name,
-                    description: data.description,
-                    license: resp.versions[0].license.clone(),
-                    documentation: data.documentation,
-                    homepage: data.homepage,
-                    repository: data.repository,
-                    total_downloads: data.downloads,
-                    max_version: data.max_version,
-                    created_at: data.created_at,
-                    updated_at: data.updated_at,
-
-                    categories: resp.categories,
-                    keywords: resp.keywords,
-                    downloads: dls,
-                    owners,
-                    reverse_dependencies,
-                    versions,
-                }
-            })
+        let name = String::from(name);
+        async move {
+            let krate = c.get_crate(&name).await?;
+            let versions = if !all_versions {
+                c.full_version(krate.versions[0].clone())
+                    .await
+                    .map(|v| vec![v])
+            } else {
+                try_join_all(
+                    krate
+                        .versions
+                        .clone()
+                        .into_iter()
+                        .map(|v| c.full_version(v)),
+                )
+                .await
+            }?;
+            let dls_fut = c.crate_downloads(&name);
+            let owners_fut = c.crate_owners(&name);
+            let reverse_dependencies_fut = c.crate_reverse_dependencies(&name);
+            try_join!(dls_fut, owners_fut, reverse_dependencies_fut).map(
+                |(dls, owners, reverse_dependencies)| {
+                    let data = krate.crate_data;
+                    FullCrate {
+                        id: data.id,
+                        name: data.name,
+                        description: data.description,
+                        license: krate.versions[0].license.clone(),
+                        documentation: data.documentation,
+                        homepage: data.homepage,
+                        repository: data.repository,
+                        total_downloads: data.downloads,
+                        max_version: data.max_version,
+                        created_at: data.created_at,
+                        updated_at: data.updated_at,
+                        categories: krate.categories,
+                        keywords: krate.keywords,
+                        downloads: dls,
+                        owners,
+                        reverse_dependencies,
+                        versions,
+                    }
+                },
+            )
+        }
     }
 
     /// Retrieve a page of crates, optionally constrained by a query.
     ///
     /// If you want to get all results without worrying about paging,
-    /// use [all_crates]().
+    /// use [`all_crates`].
     ///
     /// ```
-    pub fn crates(&self, spec: ListOptions) -> impl Future<Item = CratesResponse, Error = Error> {
+    pub fn crates(&self, spec: ListOptions) -> impl Future<Output = Result<CratesResponse>> {
         let mut url = self.base_url.join("crates").unwrap();
         {
             let mut q = url.query_pairs_mut();
@@ -259,14 +282,14 @@ impl Client {
                 q.append_pair("q", &query);
             }
         }
-        self.get(&url)
+        let c = self.clone();
+        async move { c.get(&url).await }
     }
-
     /// Retrieve all crates, optionally constrained by a query.
     ///
     /// Note: This method fetches all pages of the result.
     /// This can result in a lot queries (100 results per query).
-    pub fn all_crates(&self, query: Option<String>) -> impl Stream<Item = Crate, Error = Error> {
+    pub fn all_crates(&self, query: Option<String>) -> impl Stream<Item = Result<Crate>> {
         let opts = ListOptions {
             query,
             sort: Sort::Alphabetical,
@@ -275,23 +298,24 @@ impl Client {
         };
 
         let c = self.clone();
-        self.crates(opts.clone())
+        c.crates(opts.clone())
             .and_then(move |res| {
                 let pages = (res.meta.total as f64 / 100.0).ceil() as u64;
                 let streams_futures = (1..pages)
-                    .map(|page| {
+                    .map(move |page| {
                         let opts = ListOptions {
                             page,
                             ..opts.clone()
                         };
-                        c.crates(opts)
-                            .and_then(|res| future::ok(stream::iter_ok(res.crates)))
+                        c.crates(opts).and_then(|res| {
+                            future::ok(stream::iter(res.crates.into_iter().map(Ok)))
+                        })
                     })
                     .collect::<Vec<_>>();
-                let stream = stream::futures_ordered(streams_futures).flatten();
+                let stream = stream::FuturesOrdered::from_iter(streams_futures).try_flatten();
                 future::ok(stream)
             })
-            .flatten_stream()
+            .try_flatten_stream()
     }
 
     /// Retrieve all crates with all available extra information.
@@ -304,7 +328,7 @@ impl Client {
         &self,
         query: Option<String>,
         all_versions: bool,
-    ) -> impl Stream<Item = FullCrate, Error = Error> {
+    ) -> impl Stream<Item = Result<FullCrate>> {
         let c = self.clone();
         self.all_crates(query)
             .and_then(move |cr| c.full_crate(&cr.name, all_versions))
@@ -316,21 +340,51 @@ mod test {
     use super::*;
 
     #[test]
-    fn test_client() {
+    fn list_top_dependencies_async() -> Result<()> {
+        // Create tokio runtime
+        let mut rt = tokio::runtime::Runtime::new().unwrap();
+
+        // Instantiate the client.
+        let client = Client::new();
+        // Retrieve summary data.
+        let summary = rt.block_on(client.summary())?;
+        for c in summary.most_downloaded {
+            println!("{}:", c.id);
+            for dep in rt.block_on(client.crate_dependencies(&c.id, &c.max_version))? {
+                // Ignore optional dependencies.
+                if !dep.optional {
+                    println!("    * {} - {}", dep.id, dep.version_id);
+                }
+            }
+        }
+        Ok(())
+    }
+
+    #[test]
+    fn test_client_async() {
+        println!("Async Client test: Starting runtime");
         let mut rt = ::tokio::runtime::Runtime::new().unwrap();
 
+        println!("Creating client");
         let client = Client::new();
 
+        println!("Getting summary");
         let summary = rt.block_on(client.summary()).unwrap();
         assert!(summary.most_downloaded.len() > 0);
 
+        println!("Getting three most downloaded crates");
         for item in &summary.most_downloaded[0..3] {
-            let _ = rt.block_on(client.full_crate(&item.name, false)).unwrap();
+            println!("Geting crate: {}", &item.name);
+            let _ = rt
+                .block_on(client.full_crate(&item.name, false))
+                .unwrap();
         }
 
-        let crates = rt
-            .block_on(client.all_crates(None).take(3).collect())
-            .unwrap();
-        println!("{:?}", crates);
+        println!("Getting three crates from `all_crates`");
+        let crates = rt.block_on(client.all_crates(None).take(3).collect::<Vec<_>>());
+        for c in crates {
+            let c = c.unwrap();
+            println!("Crate: {:#?}", c);
+        }
     }
 }